feat: add images to project
