+++ package.json
@@ -124,10 +124,12 @@
   },
   "dependencies": {
     "@apollo/server": "^4.9.1",
+    "beercss": "^3.3.5",
     "eslint-plugin-import": "^2.28.0",
     "eslint-plugin-jsx-a11y": "^6.7.1",
     "eslint-plugin-react-hooks": "^4.6.0",
     "lodash": "^4.17.21",
+    "material-dynamic-colors": "^1.0.4",
     "react-imask": "^7.1.3"
   }
 }
